Ensure returned trailer length is sufficient
The srtp_get_protect_trailer_length needs to ensure
that the returned value is large enough for any of
the streams in the session.

When a session is initialized with multiple polices
it is possible to have different tag lengths for each
policy. This function provides no way to specify which
policy to use, so for now loop over all and find largest.

The current function now has limited use so suggest to make
two functions, one that takes ssrc or packet header so
correct stream can be used, second function that takes a
policy.