create single rand func for use in tests
Previously rand() was called from multiple
places in the code and has been raised as
an unsafe function call for crypto. In libsrtp
rand() has only been used for test so this change
reduces the number of locations rand() is called
to one and marks it explicitly for test usage only.